requirements.txt
accelerate==0.27.2
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
av==11.0.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
ctranslate2==4.0.0
cycler==0.12.1
distro==1.9.0
fastapi==0.110.0
faster-whisper==1.0.0
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
gradio==4.19.2
gradio_client==0.10.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
onnxruntime==1.17.1
openai==1.12.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
replicate==0.24.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
safetensors==0.4.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.36.3
streamlit==1.31.1
streamlit-audiorecorder==0.0.4
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
websockets==11.0.3
zipp==3.17.0
